requirements.txt
APScheduler==3.6.3
arrow==0.15.5
awsebcli==3.15.3
beautifulsoup4==4.8.1
blessed==1.16.1
botocore==1.12.253
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.7
django-bmemcached==0.3.0
django-bootstrap==0.2.4
django-bootstrap4==1.0.1
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-picklefield==2.1.1
django-rq==2.2.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
explicit==0.1.3
future==0.16.0
gunicorn==20.0.0
idna==2.7
importlib-metadata==1.4.0
instagram-private-api==1.6.0
instaloader==4.2.10
jmespath==0.9.4
jsonschema==2.6.0
keyring==21.1.0
kiwisolver==1.1.0
matplotlib==3.1.1
more-itertools==8.1.0
numpy==1.17.2
pathspec==0.5.9
pbr==5.4.4
Pillow==6.2.0
psycopg2==2.8.4
pyparsing==2.4.2
PyQt5==5.13.1
PyQt5-sip==4.19.19
python-binary-memcached==0.29.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==3.13
redis==3.3.11
requests==2.20.1
rq==1.2.0
selenium==3.141.0
semantic-version==2.5.0
six==1.11.0
soupsieve==1.9.5
South==1.0.2
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
tzlocal==2.0.0
uhashring==1.1
urllib3==1.24.3
virtualenv==16.7.7
wcwidth==0.1.7
websocket-client==0.56.0
whitenoise==4.1.4
zipp==2.0.1
